Added an api call example and basic page routing
