+++ src/assets/index.css
@@ -129,3 +129,32 @@ body {
   cursor: pointer;
   font-size: 48px;
 }
+
+header ul {
+  list-style-type: none;
+  border-radius: 10px;
+  margin: 20px;
+  padding: 0;
+  background-color: red;
+  overflow: hidden;
+  gap: 3px;
+}
+
+header li {
+  float: left;
+  
+
+}
+
+header a {
+  display: block;
+  color: white;
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+}
+
+header li:hover {
+  background-color: blue;
+  
+}
