+++ src/pages/MovieFinder.jsx
@@ -1,0 +1,10 @@
+import '../assets/bolt.css'
+import { useEffect, useState } from 'react'
+function MovieFinder() {
+
+    const [movies, setMovies] = useState([])
+
+
+}
+
+export default MovieFinder
